revise crond and output
Changing cron output from only a log file to a logfile and the docker default output. This removes tail -f which docker would monitor. If crond would quit docker would not know and never restart it.
This allows crond -f so it runs in the foreground and docker can monitor it.